requirements.txt
# Web Framework
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20

# NLP and AI
google-genai==1.15.0
openai-whisper==20240930
pyttsx3==2.98

# Environment and Configuration
python-dotenv==1.1.0

# HTTP and API
requests==2.32.3
httpx==0.28.1

# Data Processing
numpy==2.2.6
pydantic==2.11.4

# Utilities
typer==0.15.3
rich==14.0.0
tqdm==4.67.1
